:sparkles: Sets: Scaling and Changing Deployments
The observation is that changing the Pod template's containerPort for our "client" container in the Deployment allows us to continue seeing our React application serving up content on port 31515 (as defined by the Service). This behavior is happening because the multi-client image, which is the image that the "client" container is created from, was created with a Dockerfile that runs the EXPOSE 3000 directive. By including this in our Dockerfile, Kubernetes is allowing traffic to access that port thru the Service, as you may expect.